Donor consent and storage management

+++ services/web/app/sample/api/protocol.py
@@ -22,16 +22,14 @@ from ...misc import get_internal_api_header
 
 from ..views import new_sample_protocol_event_schema, sample_protocol_event_schema
 
-from ...database import db, SampleProtocolEvent, UserAccount, Sample, Event
-from ...database import db, SampleProtocolEvent, UserAccount
+from ...database import db, SampleProtocolEvent, UserAccount, Sample, Event, ProtocolTemplate
+from ...protocol.enums import ProtocolType
 
 
 @api.route("/sample/new/protocol_event", methods=["POST"])
 @token_required
 def sample_new_sample_protocol_event(tokenuser: UserAccount):
     values = request.get_json()
-
-    print(values)
 
     if not values:
         return no_values_response()
@@ -74,27 +72,45 @@ def sample_new_sample_protocol_event(tokenuser: UserAccount):
 @api.route("/sample/protocol_event/<uuid>/remove", methods=["POST"])
 @token_required
 def sample_remove_sample_protocol_event(uuid, tokenuser: UserAccount):
-
-    try:
-        protocol_event = SampleProtocolEvent.query.filter_by(uuid=uuid).first()
+    protocol_event = SampleProtocolEvent.query.filter_by(uuid=uuid).first()
+    if protocol_event:
         if protocol_event.is_locked:
-            err = {"messages": "Protocol Event Locked!"}
-            return validation_error_response(err)
+            return locked_response("protocol event! ")
+    else:
+        return not_found("protocol event(%s)" % uuid)
 
-        sample_uuid = Sample.query.filter_by(id=protocol_event.sample_id).first().uuid
-        print("sample_uuid: ", sample_uuid)
+    sample = Sample.query.filter_by(id=protocol_event.sample_id).\
+            with_entities(Sample.uuid, Sample.is_locked).first()
 
-    except:
-        flash("Not found")
-        return no_values_response()
+    if sample:
+        if sample.is_locked:
+            return locked_response("sample(%s)" % sample.uuid)
+    else:
+        return not_found("related sample")
+
+    sample_uuid = sample.uuid
+
+    protocol_type = ProtocolTemplate.query.filter_by(id=protocol_event.protocol_id).first().type
+    if protocol_type in [ProtocolType.ACQ, ProtocolType.ALD, ProtocolType.SDE, ProtocolType.STR]:
+        err = {"messages": "Type of protocol events (%s) not allowed!" % protocol_type}
+        return validation_error_response(err)
+
+    event = None
+    if protocol_event.event_id:
+        event = Event.query.filter_by(id=protocol_event.event_id).first()
 
     try:
         db.session.add(protocol_event)
         db.session.delete(protocol_event)
         db.session.flush()
-        db.session.commit()
+    except Exception as err:
+        return transaction_error_response(err)
 
-        flash("Sample protocol event successfully deleted!"),
+    try:
+        if event:
+            db.session.delete(event)
+
+        db.session.commit()
         return success_with_content_response(sample_uuid)
 
     except Exception as err:
