+++ services/web/app/sample/api/shipment.py
@@ -15,25 +15,29 @@
 
 
 from flask import request, abort, url_for
-from sqlalchemy.sql import func
+from sqlalchemy.sql import func, or_
 from marshmallow import ValidationError
 from ...api import api, generics
+
 import requests
 import json
 from ...api.responses import *
 from ...decorators import token_required
 from ...misc import get_internal_api_header
 from ..enums import CartSampleStorageType, SampleShipmentStatusStatus
+from .base import func_update_sample_status
+
 from ...database import (
     db,
     SampleShipmentToSample,
+    Sample,
     UserCart,
     UserAccount,
     SampleShipment,
     Event,
     EntityToStorage,
     SampleRack,
-    SampleShipmentStatus
+    SampleShipmentStatus,
 )
 
 from ..views import (
@@ -77,6 +81,7 @@ def shipment_update_status(uuid:str, tokenuser:UserAccount):
         shipment_event = SampleShipmentStatus(**new_shipment_event_values)
         shipment_event.author_id = tokenuser.id
 
+
     else:
         try:
             for attr, value in values.items():
@@ -87,13 +92,30 @@ def shipment_update_status(uuid:str, tokenuser:UserAccount):
         shipment_event.updated_on = func.now()
         shipment_event.updated_by = tokenuser.id
 
+
     try:
         db.session.add(shipment_event)
-        db.session.commit()
-        return success_with_content_response(sample_shipment_status_schema.dump(shipment_event))
+
     except Exception as err:
         return transaction_error_response(err)
 
+    sample_status_events = {"shipment_status": shipment_event}
+    res = func_update_sample_status(tokenuser=tokenuser, auto_query=True, sample=None, events=sample_status_events)
+
+    message = "Shipment status successfully updated! " + res["message"]
+
+    try:
+        if res["success"] is True and res["sample"]:
+            for sample in res["sample"]:
+                db.session.add(sample)
+
+        db.session.commit()
+        return success_with_content_message_response(
+            sample_shipment_status_schema.dump(shipment_event), message
+        )
+
+    except Exception as err:
+        return transaction_error_response(err)
 
 
 
@@ -119,7 +141,7 @@ def shipment_view_shipment(uuid: str, tokenuser: UserAccount):
                          'tracking_number': None,
                          'shipment': shipment_info,
                          'status': None}
-        print('shipment info: ', shipment_info)
+        # print('shipment info: ', shipment_info)
         return success_with_content_response(
             shipment_info
         )
@@ -147,7 +169,7 @@ def shipment_view_shipment_samples(uuid: str, tokenuser: UserAccount):
                          'tracking_number': None,
                          'shipment': shipment_info,
                          'status': None}
-        print('shipment info: ', shipment_info)
+        # print('shipment info: ', shipment_info)
         return success_with_content_response(
             shipment_info
         )
@@ -273,11 +295,11 @@ def remove_sample_from_cart(uuid: str, tokenuser: UserAccount):
             db.session.commit()
 
             return success_with_content_response(
-                {"msg": "%s removed from cart" % (uuid)}
+                {"message": "%s removed from cart" % (uuid)}
             )
         else:
             return success_with_content_response(
-                {"msg": "%s not in user cart" % (uuid)}
+                {"message": "%s not in user cart" % (uuid)}
             )
 
     else:
@@ -322,39 +344,234 @@ def add_sample_to_cart(uuid: str, tokenuser: UserAccount):
 
     if sample_response.status_code == 200:
         sample_id = sample_response.json()["content"]["id"]
+
+        sample_locked = Sample.query.filter_by(id=sample_id, is_locked=True).first()
+        if sample_locked:
+            msg_locked = "[LIMBSMP-%s]%s" % (sample_locked.id, sample_locked.uuid)
+            return locked_response(msg_locked)
+
         try:
             cart_sample_schema = new_cart_sample_schema.load({"sample_id": sample_id})
 
         except ValidationError as err:
             return validation_error_response(err)
-        check = UserCart.query.filter_by(
+
+        ESrecords = EntityToStorage.query.filter_by(sample_id=sample_id).all()
+
+        for es in ESrecords:
+            try:
+                db.session.delete(es)
+                db.session.flush()
+                # new_uc.rack_id = es.rack_id
+            except Exception as err:
+                return transaction_error_response(err)
+
+        new_uc = UserCart.query.filter_by(
             author_id=tokenuser.id, sample_id=sample_id
         ).first()
 
-        if check != None:
-            return success_with_content_response(
-                {"msg": "Sample already added to Cart"}
-            )
-
-        ESrecords = EntityToStorage.query.filter_by(sample_id=sample_id).all()
-        new_uc = UserCart(sample_id=sample_id,storage_type=None,selected=True, author_id=tokenuser.id)
-
-        for es in ESrecords:
-            db.session.delete(es)
-            db.session.flush()
-            # new_uc.rack_id = es.rack_id
+        if new_uc != None:
+            new_uc.selected = True
+            new_uc.updated_on = func.now()
+            msg = "Sample in the cart selected!"
+        else:
+            new_uc = UserCart(sample_id=sample_id, storage_type=None, selected=True, author_id=tokenuser.id)
+            msg = "Sample added to Cart!"
 
         try:
             db.session.add(new_uc)
             db.session.commit()
             db.session.flush()
 
-            return success_with_content_response({"msg": "Sample added to Cart"})
+            return success_with_content_message_response(sample_id, message=msg)
 
         except Exception as err:
             return transaction_error_response(err)
     else:
         return sample_response.content
+
+
+
+@api.route("/cart/add/samples", methods=["POST"])
+@token_required
+def add_samples_to_cart(tokenuser: UserAccount):
+    values = request.get_json()
+    samples = []
+    if values:
+        samples = values.pop('samples', [])
+
+    if len(samples) == 0:
+        return no_values_response()
+
+    sample_ids = [smpl["id"] for smpl in samples]
+
+    # Site access control
+    samples_locked = Sample.query.filter(Sample.id.in_(sample_ids), Sample.is_locked==True).\
+        with_entities(Sample.id)
+
+    if tokenuser.account_type != "Administrator":
+        samples_other = Sample.query.filter(Sample.id.in_(sample_ids), Sample.is_locked==False,
+                    Sample.current_site_id!=None, Sample.current_site_id!=tokenuser.site_id).\
+                    with_entities(Sample.id)
+
+        if samples_other.count() > 0:
+            samples_locked = samples_locked.union(samples_other)
+
+    # Samples in transit can't be added to cart
+    samples_transit = SampleShipmentStatus.query. \
+        join(SampleShipmentToSample, SampleShipmentToSample.shipment_id == SampleShipmentStatus.shipment_id). \
+        filter(~SampleShipmentStatus.status.in_(["DEL","UND", "CAN"])). \
+        with_entities(SampleShipmentToSample.sample_id)
+
+    if samples_transit.count() > 0:
+        samples_locked = samples_locked.union(samples_transit)
+
+
+    if samples_locked.count() > 0:
+        samples_locked = samples_locked.distinct().all()
+        ids_locked = [sample.id for sample in samples_locked]
+        sample_ids = [sample_id for sample_id in sample_ids if sample_id not in ids_locked]
+        msg_locked = ' | '.join(["LIMBSMP-%s" % (sample.id) for sample in samples_locked])
+    else:
+        ids_locked = []
+        msg_locked = []
+
+    if len(sample_ids) == 0:
+        return locked_response(msg_locked)
+
+    ESRecords = EntityToStorage.query.filter(EntityToStorage.sample_id.in_(sample_ids)).all()
+    n_new = 0
+    n_old = 0
+    if len(ESRecords) > 0:
+        try:
+            for es in ESRecords:
+                db.session.delete(es)
+                db.session.flush()
+        except Exception as err:
+            return transaction_error_response(err)
+
+    for sample_id in sample_ids:
+
+        new_uc = UserCart.query.filter_by(
+            author_id=tokenuser.id, sample_id=sample_id
+        ).first()
+
+        if new_uc is not None:
+            new_uc.selected = True
+            new_uc.updated_on = func.now()
+            n_old = n_old + 1
+        else:
+            new_uc = UserCart(sample_id=sample_id, selected=True, author_id=tokenuser.id)
+            n_new = n_new + 1
+
+        try:
+            db.session.add(new_uc)
+            db.session.flush()
+        except Exception as err:
+            return transaction_error_response(err)
+
+    msg = "%d samples added to Cart!" % n_new
+    if (n_old >0 ):
+        msg = msg + " | " + "%d samples updated in Cart!" % n_old
+    if len(msg_locked) >0:
+        msg = msg + " | " +"Locked sample not added: %s" % msg_locked
+
+    try:
+        db.session.commit()
+        return success_with_content_message_response(sample_ids, message=msg)
+
+    except Exception as err:
+        return transaction_error_response(err)
+
+@api.route("/cart/add/samples_racks", methods=["POST"])
+@token_required
+def add_samples_with_rack_to_cart(tokenuser: UserAccount):
+    # print("tokenuser.site_id:", tokenuser.site_id)
+    values = request.get_json()
+    samples = []
+    if values:
+        samples = values.pop('samples', [])
+
+    if len(samples) == 0:
+        return no_values_response()
+
+    sample_ids = [smpl["id"] for smpl in samples]
+
+    # -- Site access control
+    samples_locked = db.session.query(Sample.id).filter(Sample.id.in_(sample_ids), Sample.is_locked is True)
+
+    if tokenuser.account_type != "Administrator":
+        # Can only add samples of same site for the operator
+        samples_other = db.session.query(Sample.id).filter(Sample.id.in_(sample_ids), Sample.is_locked is False,
+                    Sample.current_site_id != None, Sample.current_site_id!=tokenuser.site_id)
+
+        if samples_other.count()>0:
+            samples_locked = samples_locked.union(samples_other)
+
+    # Samples in transit can't be added to cart
+    samples_transit = db.session.query(SampleShipmentToSample.sample_id). \
+        join(SampleShipmentStatus, SampleShipmentToSample.shipment_id == SampleShipmentStatus.shipment_id). \
+        filter(~SampleShipmentStatus.status.in_(["DEL","UND","CAN"]), SampleShipmentToSample.sample_id.in_(sample_ids))
+
+
+    if samples_transit.count() > 0:
+        samples_locked = samples_locked.union(samples_transit)
+
+    if samples_locked.count() >0:
+
+        ids_locked =samples_locked.all()
+        ids_locked = [i[0] for i in ids_locked]
+        sample_ids = set(sample_ids) - set(ids_locked)
+
+        msg_locked = ' | '.join(["LIMBSMP-%s" % (sample_id) for sample_id in samples_locked])
+    else:
+        ids_locked = []
+        msg_locked = []
+
+    if len(sample_ids) == 0:
+        return locked_response(msg_locked)
+
+    n_new = 0
+    n_old = 0
+    for sample_id in sample_ids:
+        new_uc = UserCart.query.filter_by(
+            author_id=tokenuser.id, sample_id=sample_id
+        ).first()
+
+        if new_uc is not None:
+            new_uc.selected = True
+            new_uc.updated_on = func.now()
+            n_old = n_old + 1
+        else:
+            new_uc = UserCart(sample_id=sample_id, selected=True, author_id=tokenuser.id)
+            n_new = n_new + 1
+
+        ets = EntityToStorage.query.filter(EntityToStorage.sample_id==sample_id,
+               EntityToStorage.rack_id is not None, EntityToStorage.removed is not True).\
+                    order_by(EntityToStorage.entry_datetime.desc()).first()
+        if ets:
+            new_uc.rack_id = ets.rack_id
+            new_uc.storage_type = "RUC"
+
+        try:
+            db.session.add(new_uc)
+            db.session.flush()
+        except Exception as err:
+            return transaction_error_response(err)
+
+    msg = "%d samples added to Cart!" % n_new
+    if (n_old >0 ):
+        msg = msg + " | " + "%d samples updated in Cart!" % n_old
+    if len(msg_locked) >0:
+        msg = msg + " | " +"Locked sample not added: %s" % msg_locked
+
+    try:
+        db.session.commit()
+        return success_with_content_message_response(sample_ids, message=msg)
+
+    except Exception as err:
+        return transaction_error_response(err)
+
 
 @api.route("/cart/add/LIMBRACK-<id>", methods=["POST"])
 @token_required
@@ -365,13 +582,19 @@ def add_rack_to_cart(id: int, tokenuser: UserAccount):
     )
 
     if rack_response.status_code == 200:
+        rackRecord = SampleRack.query.filter_by(id=id).first()
+        if not rackRecord:
+            return not_found("LIMBRACK-%d" % id)
+
+        if rackRecord.is_locked:
+            return locked_response("LIMBRACK-%d" % id)
+        rackRecord.is_locked = True
+
         esCheck = EntityToStorage.query.filter_by(rack_id=id, shelf_id=None).all()
         if esCheck == []:
-            return no_values_response()
+            return not_found("for the rack with sample(s)")
 
         ESRecords = EntityToStorage.query.filter_by(rack_id=id).all()
-        rackRecord = SampleRack.query.filter_by(id=id).first()
-        rackRecord.is_locked = True
         for es in ESRecords:
             if es.sample_id is None and es.shelf_id is not None:
                 db.session.delete(es)
@@ -388,7 +611,7 @@ def add_rack_to_cart(id: int, tokenuser: UserAccount):
 
                 if check != None:
                     return success_with_content_response(
-                        {"msg": "Sample already added to Cart"}
+                        {"message": "Sample already added to Cart"}
                     )
 
                 # es = EntityToStorage.query.filter_by(sample_id=es.sample_id).first()
@@ -399,17 +622,10 @@ def add_rack_to_cart(id: int, tokenuser: UserAccount):
                 db.session.flush()
         try:
             db.session.commit()
-            return success_with_content_response({"msg": "Sample added to Cart"})
+            return success_with_content_response({"message": "Sample added to Cart"})
 
         except Exception as err:
             return transaction_error_response(err)
-        # sample_id = sample_response.json()["content"]["id"]
-
-
-        #
-        # if not es is None:
-        #         #     new_uc.rack_id = es.rack_id
-
 
     else:
         return rack_response.content
